[sw/otbn/crypto] Add arithmetic to boolean masking for P256 ECDH
This commit modifies the p256_scalar_mult function so that it returns
an arithmetically masked result. The arithmetic masking is done before
converting back to affine coordinates. For this the internal
scalar_mult_int function is changed to return projective coordinates.
The transformation to affine coordinates now needs to be done outside
scalar_mult_int.

This commit also adds code to convert from arithmetic masking to
boolean masking. The conversion is implemented following Gaubin's
algorithm to prevent side-channel leakage. The conversion is
implemented in a 257-bit domain to enable carry bit handling,
which is necessary when the arithmetic masking is performed
within a mod p domain.

Both described changes are then put together to generate the P256
ECDH shared session key.

Further, this commit also adds the relevant tests.

Signed-off-by: Moritz Wettermann <moritz.wettermann@gi-de.com>