[sw/crypto] Fix P-256 Random Scalar Generation
This commit fixes a bug in p256_random_scalar.
To generate a random scalar, randomness is taken from RND. The WDRs which
contain the RND bits are then XORed with bits from URND, just in case
there's any vulnerability in EDN that lets the attacker recover bits
before they reach OTBN.
But one of the three WDRs that contain RND bits isn't XORed with URND.
This commit fixes this.

Signed-off-by: Moritz Wettermann <moritz.wettermann@gi-de.com>